requirements.txt
Django==3.0.4
djangorestframework==3.11.0
luigi==2.8.12
PyYAML==5.3.1
tensorflow==2.1.0
sqlalchemy==1.3.15
paho-mqtt==1.5.0
pandas==1.0.3
seaborn==0.10.0
pykafka==2.8.0
kafka-python==2.0.1
